1 - The models TipoEmpresa and TipoEvento now includes th quantity of items in each category. 2 - The TextBlocks now have the foreground white. 3 - In the App.xaml I have added a new DataTemplate to show the quantity below the title

+++ 2CantonWP/Model/Empresa.cs
@@ -19,10 +19,10 @@ namespace _2CantonWP.Model
         public string Direccion { get; set; }
           
         [JsonProperty(PropertyName = "longitud")]
-        public float Longitud { get; set; }
+        public double Longitud { get; set; }
 
         [JsonProperty(PropertyName = "latitud")]
-        public float Latitud { get; set; }
+        public double Latitud { get; set; }
 
         [JsonProperty(PropertyName = "telefonoprincipal")]
         public string TelefonoPrincipal { get; set; }
